Fixed a few warnings.
Rev: src/pike_search_engine.c:1.3
Rev: src/pike_search_engine2.c:1.2

+++ src/pike_search_engine2.c
@@ -34,7 +34,7 @@ INLINE HCHAR *NameNH(memchr_search)(void *data,
 				    HCHAR *haystack,
 				    ptrdiff_t haystacklen)
 {
-  return NameNH(MEMCHR)(haystack, (NCHAR)(long)data, haystacklen);
+  return NameNH(MEMCHR)(haystack, (NCHAR)(ptrdiff_t)data, haystacklen);
 }
